requirements.txt
fastapi~=0.115.6
pydantic
urllib3
python-multipart
sqlalchemy
crypto
pillow
websockets
opencv-python
numpy
basicsr
facexlib
realesrgan
torch
transformers~=4.47.1
uvicorn~=0.34.0
boto3~=1.35.90
botocore~=1.35.90